Separate common part for supporting index tensor
